Separate verification form implementation
